refactor : 레디스 관련 모듈 이름 변경
