+++ BE/src/middleware/session.middleware.ts
@@ -2,7 +2,7 @@ import session from 'express-session';
 import { NextFunction, Request, Response } from 'express';
 import { Injectable, NestMiddleware } from '@nestjs/common';
 import { v4 as uuidv4 } from 'uuid';
-import { CustomRedisStore } from 'src/config/redis/custom-redis-store';
+import { CustomRedisStore } from 'src/config/redisProvider/custom-redis-store';
 import { ConfigService } from '@nestjs/config';
 import { SessionRepository } from 'src/redis/session.repository';
