+++ BE/src/query/query.module.ts
@@ -6,7 +6,7 @@ import { Shell } from '../shell/shell.entity';
 import { UsageModule } from '../usage/usage.module';
 import { UserDBModule } from '../config/query-database/user-db.moudle';
 import { ShellModule } from '../shell/shell.module';
-import { RepositoriesModule } from 'src/redis/repositories.module';
+import { RedisModule } from 'src/redis/redis.module';
 
 @Module({
   imports: [
@@ -14,7 +14,7 @@ import { RepositoriesModule } from 'src/redis/repositories.module';
     UsageModule,
     UserDBModule,
     ShellModule,
-    RepositoriesModule,
+    RedisModule,
   ],
   controllers: [QueryController],
   providers: [QueryService],
