+++ BE/src/redis/redis.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
-import { RedisModule } from 'src/config/redis/redis.module';
+import { RedisProviderModule } from 'src/config/redisProvider/redis-provider.module';
 import { SessionRepository } from './session.repository';
 import { ActiveUserRepository } from './activeUser.repository';
 import { PodListRepository } from './podList.repository';
 
 @Module({
-  imports: [RedisModule],
+  imports: [RedisProviderModule],
   providers: [SessionRepository, ActiveUserRepository, PodListRepository],
   exports: [SessionRepository, ActiveUserRepository, PodListRepository],
 })
-export class RepositoriesModule {}
+export class RedisModule {}
