+++ BE/src/table/table.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { TableService } from './table.service';
 import { TableController } from './table.controller';
 import { UserDBModule } from '../config/query-database/user-db.moudle';
-import { RepositoriesModule } from 'src/redis/repositories.module';
+import { RedisModule } from 'src/redis/redis.module';
 
 @Module({
-  imports: [UserDBModule, RepositoriesModule],
+  imports: [UserDBModule, RedisModule],
   controllers: [TableController],
   providers: [TableService],
   exports: [TableService],
