+++ BE/src/usage/usage.module.ts
@@ -3,10 +3,10 @@ import { UsageController } from './usage.controller';
 import { UsageService } from './usage.service';
 import { TableModule } from '../table/table.module';
 import { UserDBModule } from '../config/query-database/user-db.moudle';
-import { RepositoriesModule } from 'src/redis/repositories.module';
+import { RedisModule } from 'src/redis/redis.module';
 
 @Module({
-  imports: [TableModule, UserDBModule, RepositoriesModule],
+  imports: [TableModule, UserDBModule, RedisModule],
   controllers: [UsageController],
   providers: [UsageService],
   exports: [UsageService],
